More reliable detection of the erc-parsed text property
Since at least Emacs 26.3, this module appears to have struggled to reliably
extract the `erc-parsed' text property in inserted chat messages during the
execution of `erc-insert-modify-hook'. This likely comes down to not expecting
the hook's members to move `point' without restoring it. Prior to this change,
`erc-colorize-message' would look for `erc-parsed' wherever `point' happened
to reside. It now favors the position returned by `erc-find-parsed-property'.

Thanks to J.P. for fixing this!